requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.3
docx-utils==0.1.2
docxtpl==0.6.3
entrypoints==0.3
enum34==1.1.6
flake8==3.7.9
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
python-dateutil==2.8.1
python-docx==0.8.7
pytz==2019.3
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
xlrd==1.2.0
